+++ lib/app_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:workoutpersonalizer_frontend/widgets/navigation_bar.dart';
+import 'package:workoutpersonalizer_frontend/widgets/navigation_bar/navigation_bar.dart';
 
 class AppView extends StatelessWidget {
   final Widget child;
