+++ lib/ui/creators_corner_page.dart
@@ -3,13 +3,32 @@ import 'package:flutter/material.dart';
 class CreatorsCornerPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Center(
-      child: Container(
-        child: const Text(
-          "Creator's Corner Page",
-          style: TextStyle(fontSize: 30),
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: <Widget>[
+        Expanded(
+          child: Row(
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: <Widget>[
+              Expanded(
+                child: Container(
+                  color: Colors.green,
+                ),
+              ),
+              Expanded(
+                child: Container(
+                  color: Colors.blue,
+                ),
+              ),
+            ],
+          ),
         ),
-      ),
+        Expanded(
+          child: Container(
+            color: Colors.red,
+          ),
+        ),
+      ],
     );
   }
 }
